hiremind/requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
pymongo==4.11.1
python-dotenv==1.0.1
bcrypt==4.2.1
PyPDF2==3.0.1
docx2txt==0.9
pdfkit==1.0.0
